move: トップページのカルーセルを `features/` に移動
