+++ frontend/src/components/features/carousels/GuideCarousel.tsx
