+++ frontend/src/components/features/carousels/GuideCarousel/Desktop/GuideItem.tsx
