+++ frontend/src/components/features/carousels/GuideCarousel/Mobile/GuideCard.tsx
