+++ frontend/src/components/features/carousels/GuideCarousel/Mobile/index.tsx
