+++ frontend/src/components/features/carousels/GuideCarousel/index.tsx
